requirements.txt
pillow==10.2.0
pytest==8.0.0
pyinstaller==6.3.0 